+++ source/stereo_transcoder_ultra/plugin.py
@@ -6,7 +6,8 @@
     Date:                     16 September 2026
 
     Copyright:
-        Unmanic plugin code Copyright (C) 2026 Eric Derewonko
+        Copyright (C) 2021 Josh.5 <jsunnex@gmail.com>
+        Copyright (C) 2026 AudioscavengeR <audioscavenger@gmail.com>
 
         This program is free software: you can redistribute it and/or modify it under the terms of the GNU General
         Public License as published by the Free Software Foundation, version 3.